refactor: consolidate IsolationModePicker and SessionTargetPicker imports in NewChatViewPane
